Struct path::Path

pub struct Path { /* fields omitted */ }

A structure that represents a file

Implementations

impl Path

pub fn new(path: String) -> Self

Creates a new Path from the given String.

pub fn components<'a>(&'a self) -> impl Iterator<Item = &'a str>

Returns an iterator over the components of this Path, split by the path delimiter "/".

pub fn rcomponents<'a>(&'a self) -> impl Iterator<Item = &'a str>

Returns a reverse iterator over the components of this Path, split by the path delimiter "/".

pub fn basename<'a>(&'a self) -> &'a str

Returns just the file name, i.e., the trailling component of the path.

Examples

"/path/to/my/file.a" -> “file.a” "my/file.a" -> “file.a” "file.a" -> “file.a”

pub fn file_stem<'a>(&'a self) -> &'a str

Like basename(), but excludes the file extension, if present.

pub fn extension<'a>(&'a self) -> Option<&'a str>

Returns the file extension, if present. If there are multiple extensions as defined by the extension delimiter, '.', then the last one will be treated as the extension.

pub fn relative(&self, other: &Path) -> Option<Path>

Returns a Path that expresses a relative path from this Path (self) to the given other Path.

pub fn is_absolute(&self) -> bool

Returns a boolean indicating whether this Path is absolute, i.e., whether it starts with the root directory.

pub fn get(&self, starting_dir: &DirRef) -> Option<FileOrDir>

Returns the file or directory specified by the given path, which can either be absolute or relative from the given starting directory.

pub fn get_file(&self, starting_dir: &DirRef) -> Option<FileRef>

Returns the file specified by the given path, which can be either absolute, or relative from the given starting directory.

If the path is invalid or points to a directory, then None is returned.

pub fn get_dir(&self, starting_dir: &DirRef) -> Option<DirRef>

Returns the file specified by the given path, which can be either absolute, or relative from the given starting directory.

If the path is invalid or points to a directory, then None is returned.

pub fn get_absolute(path: &Path) -> Option<FileOrDir>

Returns the file or directory specified by the given absolute path