Struct path::PathBuf

pub struct PathBuf { /* private fields */ }

An owned, mutable path.

This type is just a wrapper around a String.

Implementations

impl PathBuf

pub fn new() -> Self

Allocates an empty PathBuf.

pub fn push<P>(&mut self, path: P)where P: AsRef<Path>,

Extends self with path.

If path is absolute, it replaces the current path.

Examples

Pushing a relative path extends the existing path:

use std::path::PathBuf;

let mut path = PathBuf::from("/tmp");
path.push("file.bk");
assert_eq!(path, PathBuf::from("/tmp/file.bk"));

Pushing an absolute path replaces the existing path:

use std::path::PathBuf;

let mut path = PathBuf::from("/tmp");
path.push("/etc");
assert_eq!(path, PathBuf::from("/etc"));

pub fn pop(&mut self) -> bool

Truncates self to self.parent.

Returns false and does nothing if self.parent is None. Otherwise, returns true.

Examples

use std::path::{Path, PathBuf};

let mut p = PathBuf::from("/spirited/away.rs");

p.pop();
assert_eq!(Path::new("/spirited"), p);
p.pop();
assert_eq!(Path::new("/"), p);