Struct path::PathBuilder

pub struct PathBuilder {
    // some fields omitted
}

Helper to construct a Path.

Methods

impl PathBuilder

fn new() -> PathBuilder

Prepares the creation of a path. Points can be added either in order or by providing an additional parameter (which creates that order once all points were added).

fn add_point(&mut self, point: Vec4<f32>) -> &mut PathBuilder

Add points in a particular order by repeatedly calling this function.

fn add_sorted_point(&mut self, point: Vec4<f32>, param: f32) -> &mut PathBuilder

Add points by calling this function repeatedly in any order, but implicitly defining an order by providing a parameter for each point.

fn finalize(&self) -> Path

Use either points which were added in that particular order or use provided parameters to sort points added in arbitrary order.